buildutilities/requirements.txt
antlr4-python3-runtime==4.13.2
antlr4-tools==0.2.1
install-jdk==1.1.0
pyperclip==1.9.0
stringtemplate3==24.2.0
